Updated push

+++ src/main/java/com/jamub/payaccess/api/dao/Dao.java
@@ -8,7 +8,7 @@ import java.util.Optional;
 
 public interface Dao<T> {
 
-    Optional<T> get(int id);
+    Optional<T> get(Long id);
     List<T> getAll();
 //    T save(MerchantSignUpRequest t);
     void update(T t);
